Split all the option classes into their own files
