+++ spec/lib/msf/core/opt_address_range_spec.rb
