+++ spec/lib/msf/core/opt_address_spec.rb
