+++ spec/lib/msf/core/opt_bool_spec.rb
