+++ spec/lib/msf/core/opt_enum_spec.rb
