+++ spec/lib/msf/core/opt_int_spec.rb
