+++ spec/lib/msf/core/opt_path_spec.rb
