+++ spec/lib/msf/core/opt_port_spec.rb
