+++ spec/lib/msf/core/opt_raw_spec.rb
