+++ spec/lib/msf/core/opt_regexp_spec.rb
